benchmark/requirements.txt
cupy-cuda12x==13.3.0
fastrlock==0.8.2
glcm_cupy==0.2.1
imageio==2.36.1
lazy_loader==0.4
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python @ file:///home/chico/Downloads/opencv_contrib_python-4.10.0.84-cp37-abi3-linux_x86_64.whl#sha256=b5afb798b9934c248e346c707bd016cdef42ca7e65bd33542a80c3cd29bffed2
packaging==24.2
pillow==11.0.0
pydicom==3.0.1
scikit-image==0.22.0
scipy==1.14.1
tifffile==2024.9.20
tqdm==4.67.1
